requirements.txt
HTSeq
PyYAML
swalign
pyfaidx
svgwrite
regex
snakemake